producer/requirements.txt
faker==24.0.0
kafka-python-ng
pandas
numpy
scikit-learn
python-dotenv